third_party/blink/web_tests/html/menu/tentative/focus-menu-elements-arrowoperations.html

<!DOCTYPE html>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>

<menubar>
 <menuitem id=btn commandfor="list" command="toggle-popover">Open</menuitem>
</menubar>

<menulist id="list">
 <menuitem id="A1">Command A1</menuitem>
 <menuitem id="A2">Command A2</menuitem>
</menulist>

<menubar id="bar">
  <menuitem id="B1">Command B1</menuitem>
  <menuitem id="B2">Command B2</menuitem>
</menubar>


<script>

test(() => {
  if (!window.eventSender)
    return;
  eventSender.keyDown('Tab');
  assert_equals(document.activeElement, btn);
  eventSender.keyDown('Enter');
  assert_equals(document.activeElement, btn, 'btn invoked menulist, but focus is still on btn.');
  eventSender.keyDown('ArrowDown');
  assert_equals(document.activeElement, A1, 'arrow down moves focus into menulist.');
  eventSender.keyDown('ArrowLeft');
  assert_equals(document.activeElement, A1, 'arrow left does not change current focused element.');
  eventSender.keyDown('ArrowRight');
  assert_equals(document.activeElement, A1, 'arrow right does not change current focused element.');
  eventSender.keyDown('ArrowDown');
  assert_equals(document.activeElement, A2, 'arrow down changes to next menuitem.');
  eventSender.keyDown('ArrowUp');
  assert_equals(document.activeElement, A1, 'arrow up changes to previous menuitem.');
  eventSender.keyDown('ArrowUp');
  assert_equals(document.activeElement, A1, 'arrow keys do not loop.');
}, 'Should use arrow keys to move between menuitems in menulist.');

test(() => {
  if (!window.eventSender)
    return;
  B1.focus();
  assert_equals(document.activeElement, B1);
  eventSender.keyDown('ArrowUp');
  assert_equals(document.activeElement, B1, 'arrow up does not change current focused element.');
  eventSender.keyDown('ArrowDown');
  assert_equals(document.activeElement, B1, 'arrow down does not change current focused element.');
  eventSender.keyDown('ArrowRight');
  assert_equals(document.activeElement, B2, 'arrow right changes to next menuitem.');
  eventSender.keyDown('ArrowLeft');
  assert_equals(document.activeElement, B1, 'arrow left changes to previous menuitem.');
  eventSender.keyDown('ArrowLeft');
  assert_equals(document.activeElement, B1, 'arrow keys do not loop.');
}, 'Should use arrow keys to move between menuitems in menubar.');

</script>
